Calculate Effective Reproduction Number from Covidestim output

Creates a data.frame with Rt estimates over time and 95% CIs.

# S3 method for covidestim_result
RtEst(ccr, window.length = 5, mean.si = 4.7, std.si = 2.9)

# S3 method for covidestim_result
RtNaiveEstim(cc, window.length = 5, mean.si = 4.7, std.si = 2.9)

Arguments

ccr
The result of calling run. An object of class covidestim_result.

window.length
The size of the moving window over which Rt should be estimated; defaults to 5 days. This number must be odd, so that the interval's true center always falls on a complete day. It also must be less than half the number of data days, i.e., window.length < ndays/2.

mean.si
The mean serial interval to use in the Rt estimate; defaults to 4.7 days.

std.si
The standard deviation of the serial interval; defaults to 2.9 days.

Value

A data.frame with columns date, Rt, Rt.lo, Rt.hi, where date is a vector of Date objects. Or, for RtNaiveEstim, columns date, NaiveRt, NaiveRt.lo, NaiveRt.hi

Details

This code makes use of the EpiEstim package, described here:

Cori A, Ferguson NM, Fraser C, Cauchemez S (2013). “A New Framework and Software to Estimate Time-Varying Reproduction Numbers During Epidemics.” American Journal of Epidemiology, 178(9), 1505--1512. ISSN 1476-6256, 0002-9262, doi: 10.1093/aje/kwt133 , https://academic.oup.com/aje/article-lookup/doi/10.1093/aje/kwt133.

The method used in EpiEstim is "parametric_si", and a 5-day sliding window is used by default. Warning messages specifying that you are "estimating R too early" are expected and should be ignored.