Fixed values on delay to progression

This function returns a keyed list of priors related to progression to the infectious state. Called with no arguments, the default values are returned. Custom hyperpriors can be specified by passing values for the parameters specified below. The return value of this function must be added to a covidestim object using the addition operator (see examples). The following arguments can be passed to create different priors:

priors_progression(
  inf_prg_delay = c(3.413, 0.6051 * 7),
  sym_prg_delay = c(1.624, 0.2175 * 7),
  sev_prg_delay = c(2.061, 0.2277 * 7),
  asy_rec_delay = c(14, 2 * 7),
  pri_serial_i = c(34.615, 11.538 * 7),
  infect_dist = c(8, 1.241 * 7),
  seropos_dist = c(4.41, 0.042 * 7)
)

Arguments

inf_prg_delay

A two-element numeric vector containing c(shape, rate) parameters/hyperpriors of a Gamma distribution modeling the time from infection to sypmtom onset.

Source for default value: Lauer SA, Grantz KH, Bi Q, Jones FK, Zheng Q, Meredith HR, Azman AS, Reich NG, Lessler J (2020). “The Incubation Period of Coronavirus Disease 2019 (COVID-19) From Publicly Reported Confirmed Cases: Estimation and Application.” Annals of Internal Medicine, 172(9), 577--582. ISSN 0003-4819, 1539-3704, doi:10.7326/M20-0504 , https://www.acpjournals.org/doi/10.7326/M20-0504.

sym_prg_delay

A two-element numeric vector containing c(shape, rate) parameters/hyperpriors of a Gamma distribution modeling the time (in number of days) from symptom onset to severe disease.

Source for default value: Zhou F, Yu T, Du R, Fan G, Liu Y, Liu Z, Xiang J, Wang Y, Song B, Gu X, Guan L, Wei Y, Li H, Wu X, Xu J, Tu S, Zhang Y, Chen H, Cao B (2020). “Clinical course and risk factors for mortality of adult inpatients with COVID-19 in Wuhan, China: a retrospective cohort study.” The Lancet, 395(10229), 1054--1062. ISSN 01406736, doi:10.1016/S0140-6736(20)30566-3 , https://linkinghub.elsevier.com/retrieve/pii/S0140673620305663.

sev_prg_delay

A two-element numeric vector containing c(shape, rate) parameters/hyperpriors of a Gamma distribution modeling the time (in number of days) from onset of severe symptoms to death.

Source for default value: Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S, Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other Epidemiological Characteristics of 2019 Novel Coronavirus Infections with Right Truncation: A Statistical Analysis of Publicly Available Case Data.” Journal of Clinical Medicine, 9(2), 538. ISSN 2077-0383, doi:10.3390/jcm9020538 , https://www.mdpi.com/2077-0383/9/2/538.

asy_rec_delay

A two-element numeric vector containing c(shape,rate) parameters/hyperpriors of a Gamma distribution modeling the time (in number of days) from infection to recovery for individuals who never develop symptoms.

pri_serial_i

A two-element numeric vector containing c(shape,rate) parameters/hyperpriors of a Gamma distribution modeling the serial interval, the average time (in number of days) between successive cases.

infect_dist

A two-element numeric vector containing c(shape,rate) parameters/hyperpriors of a Gamma distribution modeling the changes of infectiousness, following initial infection. The PDF of this distribution can be thought of as describing how infectious an individual is, as a function of time (in days).

seropos_dist

A two-element numeric vector containing c(shape,rate) parameters/hyperpriors of a Gamma distribution modeling the time (in days) to seroreversion, following initial infection. We do not recommend modifying this value, as it is an experimental feature, and not used in the core model.

Value

An S3 object of class priors

Examples

cfg <- covidestim(ndays = 50, region = 'New York') + priors_progression(inf_prg_delay = c(4, 1))
#> Error in covidestim(ndays = 50, region = "New York"): unused argument (ndays = 50)