Priors for transitions

This function returns a keyed list of priors related to progression to various health states. Called with no arguments, the default values are returned. All parameters must be non-negative real numbers.

priors_transitions(
  p_sym_if_inf = c(44.9, 9.9),
  p_sev_if_sym = c(11.9, 47.7),
  p_die_if_sev = c(3, 97),
  p_die_if_sym = c(1.12, 85.1)
)

Arguments

p_sym_if_inf
A two-element numeric vector containing c(alpha, beta) parameters of a Beta distribution modeling the probability of being symptomatic if infectious.

Sources for default value:

Byambasuren O, Cardona M, Bell K, Clark J, McLaws M, Glasziou P (2020). “Estimating the extent of asymptomatic COVID-19 and its potential for community transmission: systematic review and meta-analysis.” Infectious Diseases (except HIV/AIDS). doi: 10.1101/2020.05.10.20097543 , http://medrxiv.org/lookup/doi/10.1101/2020.05.10.20097543.

Mizumoto K, Kagaya K, Zarebski A, Chowell G (2020). “Estimating the asymptomatic proportion of coronavirus disease 2019 (COVID-19) cases on board the Diamond Princess cruise ship, Yokohama, Japan, 2020.” Eurosurveillance, 25(10). ISSN 1560-7917, doi: 10.2807/1560-7917.ES.2020.25.10.2000180 , https://www.eurosurveillance.org/content/10.2807/1560-7917.ES.2020.25.10.2000180.

Nishiura H, Kobayashi T, Miyama T, Suzuki A, Jung S, Hayashi K, Kinoshita R, Yang Y, Yuan B, Akhmetzhanov AR, Linton NM (2020). “Estimation of the asymptomatic ratio of novel coronavirus infections (COVID-19).” International Journal of Infectious Diseases, 94, 154--155. ISSN 12019712, doi: 10.1016/j.ijid.2020.03.020 , https://linkinghub.elsevier.com/retrieve/pii/S1201971220301399.

p_sev_if_sym
A two-element numeric vector containing c(alpha, beta) parameters of a Beta distribution modeling the probability of severe if symptomatic.

Sources for default value:

CDC COVID-19 Response Team, CDC COVID-19 Response Team, Bialek S, Boundy E, Bowen V, Chow N, Cohn A, Dowling N, Ellington S, Gierke R, Hall A, MacNeil J, Patel P, Peacock G, Pilishvili T, Razzaghi H, Reed N, Ritchey M, Sauber-Schatz E (2020). “Severe Outcomes Among Patients with Coronavirus Disease 2019 (COVID-19) — United States, February 12–March 16, 2020.” MMWR. Morbidity and Mortality Weekly Report, 69(12), 343--346. ISSN 0149-2195, 1545-861X, doi: 10.15585/mmwr.mm6912e2 , http://www.cdc.gov/mmwr/volumes/69/wr/mm6912e2.htm?s_cid=mm6912e2_w.

Verity R, Okell LC, Dorigatti I, Winskill P, Whittaker C, Imai N, Cuomo-Dannenburg G, Thompson H, Walker PGT, Fu H, Dighe A, Griffin JT, Baguelin M, Bhatia S, Boonyasiri A, Cori A, Cucunubá Z, FitzJohn R, Gaythorpe K, Green W, Hamlet A, Hinsley W, Laydon D, Nedjati-Gilani G, Riley S, van Elsland S, Volz E, Wang H, Wang Y, Xi X, Donnelly CA, Ghani AC, Ferguson NM (2020). “Estimates of the severity of coronavirus disease 2019: a model-based analysis.” The Lancet Infectious Diseases, 20(6), 669--677. ISSN 14733099, doi: 10.1016/S1473-3099(20)30243-7 , https://linkinghub.elsevier.com/retrieve/pii/S1473309920302437.

p_die_if_sev
A two-element numeric vector containing c(alpha, beta) parameters of a Beta distribution modeling the probability of dying if severely ill.

Source for default value: CDC COVID-19 Response Team, CDC COVID-19 Response Team, Bialek S, Boundy E, Bowen V, Chow N, Cohn A, Dowling N, Ellington S, Gierke R, Hall A, MacNeil J, Patel P, Peacock G, Pilishvili T, Razzaghi H, Reed N, Ritchey M, Sauber-Schatz E (2020). “Severe Outcomes Among Patients with Coronavirus Disease 2019 (COVID-19) — United States, February 12–March 16, 2020.” MMWR. Morbidity and Mortality Weekly Report, 69(12), 343--346. ISSN 0149-2195, 1545-861X, doi: 10.15585/mmwr.mm6912e2 , http://www.cdc.gov/mmwr/volumes/69/wr/mm6912e2.htm?s_cid=mm6912e2_w.

p_die_if_sym
A two-element numeric vector containing c(alpha, beta) parameters of a Beta distribution modeling the probability of dying if symptomatic.

Source for default value: Basu A (2020). “Estimating The Infection Fatality Rate Among Symptomatic COVID-19 Cases In The United States: Study estimates the COVID-19 infection fatality rate at the US county level.” Health Affairs, 10.1377/hlthaff. ISSN 0278-2715, 1544-5208, doi: 10.1377/hlthaff.2020.00455 , http://www.healthaffairs.org/doi/10.1377/hlthaff.2020.00455.

Value

An S3 object of class priors

Examples

cfg <- covidestim(ndays = 50) + priors_transitions(p_sym_if_inf = c(0.5, 0.2))