Summarize a Covidestim run

Returns a data.frame summarizing a Covidestim model run. Note that if runOptimizer.covidestim is used, all *_p(2_5|25|75|97_5) variables will be NA-valued: the optimizer does not produce intervals.

# S3 method for covidestim_result
summary(ccr, include.before = TRUE, index = FALSE)

Arguments

ccr

A covidestim_result object

include.before

A logical scalar. Include estimations that fall in the period before the first day of input data? (This period is of length nweeks_before as passed to covidestim). If TRUE, any elements of variables which do not have values for this "before" period will be represented as NA.

index

A logical scalar. If TRUE, will include a variable index in the output, with range 1:(nweeks_before + nweeks).

Value

A data.frame with the following variables:

date

Date as a Date vector.

deaths, deaths + (_p2_5, _p25, _p75, _p97_5)

The number deaths estimated to occur on date date.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

deaths_of_diagnosed, deaths_of_diagnosed + (_p2_5, _p25, _p75, _p97_5)

The number of diagnosed deaths estimated to occur on date date.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

diagnoses, diagnoses + (_p2_5, _p25, _p75, _p97_5)

The number of diagnoses estimated to occur on date date.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

diagnoses_of_symptomatic, diagnoses_of_symptomatic + (_p2_5, _p25, _p75, _p97_5)

The number of diagnoses of symptomatic individuals estimated to occur on date date.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

effective_protection_inf_prvl, effective_protection_inf_prvl + (_p2_5, _p25, _p75, _p97_5)

The estimated fraction of the population on date date with effective protection against infection with a history of infection (but no vaccinations).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

effective_protection_inf_vax_boost_prvl, effective_protection_inf_vax_boost_prvl + (_p2_5, _p25, _p75, _p97_5)

The estimated fraction of the population on date date with effective protection against infection, with a history of infection, vaccination and a booster shot.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

effective_protection_inf_vax_prvl, effective_protection_inf_vax_prvl + (_p2_5, _p25, _p75, _p97_5)

The estimated fraction of the population on date date with effective protection against infection, with a history of infection and vaccination (but no booster shot).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

effective_protection_vax_boost_prvl, effective_protection_vax_boost_prvl + (_p2_5, _p25, _p75, _p97_5)

The estimated fraction of the population on date date with effective protection against infection with a history of vaccination and a booster shot (but no infection).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

effective_protection_vax_prvl, effective_protection_vax_prvl + (_p2_5, _p25, _p75, _p97_5)

The estimated fraction of the population on date date with effective protection against infection with a history of vaccination (but no booster shot or infection).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

fitted_cases, fitted_cases + (_p2_5, _p25, _p75, _p97_5)

The number of modeled case reports for a date date. This takes the delay from diagnosis to report into account and thus is approximating how many case reports should exist for this date. This estimate is used to fit against the observed data (reported diagnoses).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

fitted_deaths, fitted_deaths + (_p2_5, _p25, _p75, _p97_5)

The number of modeled death reports for a date date. This takes the delay from death to report into account and thus is approximating how many death reports should exist for this date. This estimate is used to fit against the observed data (reported deaths).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

fitted_hospitalizations, fitted_hospitalizations + (_p2_5, _p25, _p75, _p97_5)

The number of modeled hospitalization reports for a date date. This takes the delay from hospitalization to report into account and thus is approximating how many admission reports should exist for this date. This estimate is used to fit against the observed data (reported hospital admissions).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

fitted_wastewater_prvl, fitted_wastewater_prvl + (_p2_5, _p25, _p75, _p97_5)

To be developed - a modeled estimate that resembles the wastewater data; a measure of infectiousness in the population on date date.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

immunoexposed_cumulative, immunoexposed_cumulative + (_p2_5, _p25, _p75, _p97_5)

The estimated fraction of the population on date date with historic immunological exposure (infection and/or vaccination).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

infections_cumulative, infections_cumulative + (_p2_5, _p25, _p75, _p97_5)

The estimated cumulative number of infections on date date. This includes both first and repeat infections and can therefore exceed the population size.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

infections, infections + (_p2_5, _p25, _p75, _p97_5)

The number of infections estimated to occur on date date. This includes both first and repeat infections.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

infections_premiere, infections_premiere + (_p2_5, _p25, _p75, _p97_5)

The number of first infections estimated to occur on date date.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

r_t, r_t + (_p2_5, _p25, _p75, _p97_5)

Estimate of the effective reproductive number (\(R_t\)).

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

seropositive_prvl, seropositive_prvl + (_p2_5, _p25, _p75, _p97_5)

The number of individuals in the population who are modeled as being seropositive. This is an unreliable outcome that we don't recommend using.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

severe, severe + (_p2_5, _p25, _p75, _p97_5)

The number of transitions into the "severe" health state on date date. The "severe" state is defined as disease that would merit hospitalization.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

susceptible_prvl, susceptible_prvl + (_p2_5, _p25, _p75, _p97_5)

The fraction of the population on date date that is susceptible to SARS-CoV-2 infection, i.e., the fraction of the population that has no effective protection.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

susceptible_severe_prvl, susceptible_severe_prvl + (_p2_5, _p25, _p75, _p97_5)

The fraction of the population on date date that is susceptible to developing severe disease from a SARS-CoV-2 infection. This is a subset from susceptible_prvl.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

symptomatic, symptomatic + (_p2_5, _p25, _p75, _p97_5)

The number of modeled transitions of infected individuals into the infected, symptomatic health state on date date. This takes into account the probability of becoming symptomatic and the delay between infection and presentation of symptoms.

Median, 2.5% interval, 25% interval, 75% interval, 97.5% interval, ℝ.

data_available

Was there input data available for date date? This should be TRUE, except for the first month of estimates.

logical.