Input observational data

A family of functions used for inputting data into Covidestim.

input_cases(data, type = "reported")

input_deaths(data, type = "reported")

input_fracpos(data)

Arguments

data
A data.frame as described below.

type
A string, either "reported" or "occurred".

"reported" applies to the following situations:

Case or death data where the count for a particular day represents the number of cases or deaths that were reported publicly on that day (for example, as of June 1st, 2020, this is the type of data reported by the New York Times' covid-19-data repository, and by the Covid Tracking Project.

Cases or deaths data where the count for a particular day is the number of cases or deaths received from hospital reports, or diagnostic laboratories, accumulated across a particular region

"occurred" applies to the following situations:

Data where counts represent the number of of cases or deaths that actually occured on that day. For case data, this would be the day a test was administered or the date of symptom onset. For deaths, this would be the day an individual died. Note that, if the deaths data represented the day an individual's death was first reported as a SARS-Cov-2-related death, that data should be passed with type = "reported", instead.

Details

There are three types of observational data that can be used with Covidestim:

Case data, detailing the number of new cases each day

Death data, detailing the number of confirmed Covid-19 deaths each day

Testing data, detailing the fraction of positive tests for each day of data.

All input data to Covidestim is expected to be a data.frame of two variables. One variable, date must be a vector of type POSIXct or Date. The second column, observations will be a non-negative numeric vector, except in the case of testing data, where the column must be a numeric vector between [0,1].

Fraction positve data should represent, for a given day, the number of positive test results reported that day, over the total number of test results received that day. Note that the model tends to perform poorly on fraction positive data that contains dramatic jumps. With highly variable data, we recommened users smooth the fraction positive data with a moving average,for example by using the function sma. Reasons for this variation, as they have manifested in publicly available datasets, have included:

Dumping of test results: Some states have a tendency to dump batches of results on particular days.

Accumulation of test results reported over the weekend: Many states have lower reporting volume over the weekend, creating an unstable fraction positive quantity as the sample size increases and decreases dramatically.

Reclassification of results: Some states have retroactively changed test results, or removed test results entirely, and these changes may manifest on a single day, rather than being applied to all previous days affected by the change.

Antibody tests: Some states have been pooling antibody tests with viral tests, as reported in The Atlantic, and are now removing that pooling from their data.

Missing values in cases and deaths data should be represented as 0. The date range of all sets of data passed to covidestim must be equivalent, with one observation each day, and no gaps in the data. Assertions attempt to enforce this specification.