Summarize a Covidestim run

Returns a data.frame summarizing a Covidestim model run. Note that if runOptimizer.covidestim is used, all *.(lo|hi) variables will be NA-valued, because BFGS does not generate confidence intervals.

# S3 method for covidestim_result
summary(ccr, include.before = TRUE, index = FALSE)

Arguments

ccr

A covidestim_result object

include.before

A logical scalar. Include estimations that fall in the period before the first day of input data? (This period is of length ndays_before as passed to covidestim). If TRUE, any elements of variables which do not have values for this "before" period will be represented as NA.

index

A logical scalar. If TRUE, will include a variable index in the output, with range 1:(ndays_before + ndays).

Value

A data.frame with the following variables:

date

Date as a Date vector.

Rt, Rt.lo, Rt.hi

Estimate of the effective reproductive number (\(R_t\)). Median and 95% interval, ℝ.

infections, infections.lo, infections.hi

The number of modeled infections that occurred on date date. This includes infections that may never cause symptoms, as well as infections which will never show up in case reports (will never be diagnosed). Being indexed by date-of-occurrence, reporting lag is absent from this outcome. Median and 95% interval, ℝ.

cum.incidence, cum.incidence.lo, cum.incidence.hi

The number of modeled cumulative infections (not cases, or diagnoses) that have occurred by the end of date date. Median and 95% interval, ℝ.

diagnoses, diagnoses.lo, diagnoses.hi

The number of modeled diagnoses that occurred on date date. This is the sum of:

New asmptomatic diagnoses on date date, plus:

New diagnoses of symptomatic, non-severe individuals on date date, plus:

New diagnoses of severe individuals on date date.

Median and 95% interval, ℝ.

cases.fitted, cases.fitted.lo, cases.fitted.hi

The number of modeled case reports for date date. This will always differ from the number of observed cases: it is the model's approximation of how many case reports should have been filed that day. This outcome reflects underascertainment, and the full delay structure downstream of infection events. Note: infection events are tabulated by the infections[.lo/.hi] outcomes. Median and 95% interval, ℝ.

symptomatic, symptomatic.lo, symptomatic.hi

The number of modeled transitions of infected individuals into the infected, symptomatic health state on date date. This takes into account the probability of becoming symptomatic and the delay between infection and presentation of symptoms. Median and 95% interval, ℝ.

symptomatic.diagnosed, symptomatic.diagnosed.lo, symptomatic.diagnosed.hi

The number of modeled diagnoses of symptomatic individuals occurring on date date. The difference between this outcome and the diagnoses outcome is that diagnoses includes modeled diagnoses of asymptomatic individuals. Median and 95% interval, ℝ.

severe, severe.lo, severe.hi

The number of transitions into the "severe" health state on date date. The "severe" state is defined as disease that would merit hospitalization. This outcome is not intended to model observational data detailing the number of COVID-positive hospital admissions or COVID-primary-cause hospital admissions. Median and 95% interval, ℝ.

deaths, deaths.lo, deaths.hi

The number of modeled deaths for date date. The number of deaths estimated to have occurred on date date and does not account for reporting delays. Median and 95% interval, ℝ.

deaths.fitted, deaths.fitted.lo, deaths.fitted.hi

The number of modeled death reports for a date date. This will always differ from the number of observed deaths for that same date, because deaths.fitted is approximating how many death reports should exist for that date. Median and 95% interval, ℝ.

data.available

Was there input data available for date date? This should be TRUE, except for the first month of estimates. logical.

sero.positive, sero.positive.lo, sero.positive.hi

The number of individuals in the population who are modeled as being seropositive. This is an unreliable outcome that we don't recommend using. Median and 95% interval, ℝ.

pop.infectiousness, pop.infectiousness.lo, pop.infectiousness.hi

Meant to be an estimate of infectiousness of the population for comparison with wastewater data. This is an unreliable outcome that we don't recommend using. Median and 95% interval, ℝ.