Summarize a Covidestim run

Summarize a Covidestim run

# S3 method for covidestim_result
summary(ccr, include.before = TRUE, index = FALSE)

Arguments

ccr
A covidestim_result object

include.before
A logical scalar. Include estimations that fall in the period before the first day of input data? (This period is of length ndays_before as passed to covidestim). If TRUE, any elements of variables which do not have values for this "before" period will be represented as NA.

index
A logical scalar. If TRUE, will include a variable index in the output, with range 1:(ndays_before + ndays).

Value

A data.frame with the following variables:

date

cases.fitted, cases.fitted.lo, cases.fitted.hi: Median and 95% interval around fitted case data.

cum.incidence, cum.incidence.lo, cum.incidence.hi: Median and 95% interval around cumulative incidence data.

deaths.fitted, deaths.fitted.lo, deaths.fitted.hi: Median and 95% interval around fitted deaths data.

deaths, deaths.lo, deaths.hi: Median and 95% interval around estimated deaths/day.

infections, infections.lo, infections.hi: Median and 95% interval around estimated infections/day.

severe, severe.lo, severe.hi: Median and 95% interval around "severe" estimate.

symptomatic, symptomatic.lo, symptomatic.hi: Median and 95% interval around estimate of quantity of individuals symtomatic on a particular day.

data.available: TRUE/FALSE for whether input data was available on that particular day.

Rt, Rt.lo, Rt.hi: Estimate of the effective reproductive number (Rt), with 95% CIs.

sero.positive, sero.positive.lo, sero.positive.hi: Estimate of the number of individuals sero- positive, with 95% CIs.

pop.infectiousness, pop.infectiousness.lo, pop.infectiousness.hi: Estimate of the relative level of viral shedding in the community, with 95% CIs.

index